Add a NUM-EPISODES argument to LEARN
Since <RLM> already provides this argument to specify the number of episodes
it seems to be a good idea to also provide it for LEARN.  What would be even
more useful would be an additional argument MAX-STEPS-PER-EPISODE for
cases where episodes can become arbitrarily bad, e.g., when a wrong decision
is taken at the top-level.  Th same effect can currently be achieved by signaling
ENV-TERMINATED, but this solution is not particularly elegant.